refactor: 도메인 로직은 domain 하위로, ui관련 로직은 view 하위에서 관리한다
